auth

+++ Pages/Animals/Delete.cshtml.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +12,8 @@ using PetAdopt.Models;
 
 namespace PetAdopt.Pages.Animals
 {
+    [Authorize(Policy = "EditOrDeletePolicy")]
+
     public class DeleteModel : PageModel
     {
         private readonly PetAdopt.Data.PetAdoptContext _context;
@@ -20,22 +24,24 @@ namespace PetAdopt.Pages.Animals
         }
 
         [BindProperty]
-      public Animal Animal { get; set; } = default!;
+        public Animal Animal { get; set; } = default!;
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
-            if (id == null || _context.Animal == null)
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (id == null || userId == null || _context.Animal == null)
             {
                 return NotFound();
             }
 
-            var animal = await _context.Animal.FirstOrDefaultAsync(m => m.id == id);
+            var animal = await _context.Animal.FirstOrDefaultAsync(m => m.id == id && m.UserId == userId);
 
             if (animal == null)
             {
                 return NotFound();
             }
-            else 
+            else
             {
                 Animal = animal;
             }
@@ -44,13 +50,16 @@ namespace PetAdopt.Pages.Animals
 
         public async Task<IActionResult> OnPostAsync(int? id)
         {
-            if (id == null || _context.Animal == null)
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (id == null || userId == null || _context.Animal == null)
             {
                 return NotFound();
             }
+
             var animal = await _context.Animal.FindAsync(id);
 
-            if (animal != null)
+            if (animal != null && animal.UserId == userId)
             {
                 Animal = animal;
                 _context.Animal.Remove(Animal);
